agents_requirements.txt
crewai>=0.30.0
python-dotenv>=1.0.0
uuid>=1.30.0
langchain>=0.1.0 